get image annotations

Calls the 'Google Vision' API to return annotations. The function automatically creates batches

get_annotations(images, features, max_res, mode)

Arguments

images
A character vector for images to be annotated. Can either be url strings or local images, as specified with mode.

features
A character vector for the features to be returned. Accepts 'all' or any combination of the following inputs: 'label', 'web', 'text', 'face', 'landmark', 'logo', 'safe_search', 'object', 'properties'

max_res
An integer specifying the maximum number of results to be returned for each feature.

mode
Accepts 'url' for image urls and 'local' for file paths to local images.

Value

An response object of class 'gvision_annotations'.

See also

Examples

 gvision_init()

#> Succesfully initialized authentification credentials.

 # one image url
 sw_image <- 'https://upload.wikimedia.org/wikipedia/en/a/a2/Star_Wars_The_Force_Awakens_Theatrical_Poster.jpg'
 results <- get_annotations(images = sw_image, # image character vector
                           features = 'all', # request all available features
                           max_res = 10, # maximum number of results per feature
                           mode = 'url')  # maximum number of results per feature

#> [1] "Sending API request(s).."

 # multiple image urls
 finn_image <- 'https://upload.wikimedia.org/wikipedia/en/2/2a/Finn-Force_Awakens_%282015%29.png'
 padme_image <- 'https://upload.wikimedia.org/wikipedia/en/e/ee/Amidala.png'

 input_imgs <- c(sw_image, finn_image, padme_image)
 results <- get_annotations(images = input_imgs,
            features = c('labels', 'faces'), max_res = 5, mode = 'url')

#> Error in get_annotations(images = input_imgs, features = c("labels", "faces"),     max_res = 5, mode = "url"): invalid "features" provided. Check the documentation for valid features.

 # one local image
 temp_img_path <- tempfile(fileext = '.png')
 download.file(finn_image, temp_img_path, mode = 'wb', quiet = TRUE)

 results <- get_annotations(images = temp_img_path,
            features = c('labels', 'faces'), max_res = 5, mode = 'local')

#> Error in get_annotations(images = temp_img_path, features = c("labels",     "faces"), max_res = 5, mode = "local"): invalid "features" provided. Check the documentation for valid features.